moving Geography into models module; trimming api keys
