+++ src/bbd/__init__.py
@@ -1,8 +1,15 @@
 __version__ = "0.0.7"
 
-# TODO add relevant imports...
-from .working_directory import working_directory
-from .geocoder import geocoder
+from . import working_directory
+# import bbd.geocoder as geocoder
+# import bbd.census as census
+# import bbd.fec as fec
+# import bbd.gis as gis
 
+from . import geocoder
+from . import census
+from . import fec
+from . import gis
+from . import models
 
-__all__ = [working_directory, geocoder]
+__all__ = [working_directory, geocoder, census, fec, gis, models]
