+++ src/bbd/census/__init__.py
@@ -1,5 +1,4 @@
 from .get_shapefile import get_shapefile
-from .geography import Geography
 from .datasets import DataSets
 from .load import load_json_file, load_json_str
 from .get_acs import get_acs, construct_api_call
@@ -7,7 +6,6 @@ from .api_key import api_key
 
 __all__ = [
     get_shapefile,
-    Geography,
     DataSets,
     load_json_file,
     load_json_str,
