--- src/bbd/census/geography.py
@@ -1,10 +1,0 @@
-class Geography:
-    """Geography available for download"""
-
-    TRACT = "tract"
-    CD = "congressional district"
-    COUNTY = "county"
-    STATE = "state"
-    ZCTA = "zip code tabulation area"
-    BLOCK = "block"
-    BLOCKGROUP = "block group"
