+++ src/bbd/fec/api_key.py
@@ -17,6 +17,6 @@ class _ApiKey:
                 "Value should be a 'str.'"
             )
         else:
-            self._key = key_value
+            self._key = key_value.strip()
 
 api_key = _ApiKey()
