+++ src/bbd/models/__init__.py
@@ -1,0 +1,3 @@
+from .geography import Geography
+
+__all__ = [Geography]
